Refactor Key Promoter X and made it a dynamic plugin
Major refactor of KeyPromoter.java and its functions.
Split the handling of Platform actions and raw AWT events.
Now these classes are instantiated separately using the Platform's listener and services.

With this refactor, the plugin should be finally a dynamic plugin which does not require a restart when updating.

Update the build system and fixed several places where services were not acquired dynamically.